requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
blinker==1.9.0
boto3==1.39.14
botocore==1.39.14
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
cryptography==45.0.5
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.3
django-deep-serializer==0.1.3
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
Flask==3.1.0
idna==3.10
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mysql-connector-python==9.3.0
nest-asyncio==1.6.0
numpy==2.2.6
packaging==23.2
pandas==2.3.1
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.8.2
pytz==2025.2
pywin32==306
pyzmq==25.1.2
requests==2.32.4
s3transfer==0.13.1
six==1.16.0
snowflake-connector-python==3.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.3
tomlkit==0.13.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.5.0
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
Werkzeug==3.1.3
